rename: rename StationMarkersContainer to HighZoomMarkerContainer
[#811]
