+++ frontend/src/components/google-maps/map/CarFfeineListener.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import { useQueryClient } from '@tanstack/react-query';
 
-import { useRenderStationMarker } from '@marker/StationMarkersContainer/hooks/useRenderStationMarker';
+import { useRenderStationMarker } from '@marker/HighZoomMarkerContainer/hooks/useRenderStationMarker';
 
 import { debounce } from '@utils/debounce';
 import { useExternalValue, useSetExternalState } from '@utils/external-state';
