+++ frontend/src/components/google-maps/marker/HighZoomMarkerContainer/components/CarFfeineMarker/CarFfeine.stories.tsx
