+++ frontend/src/components/google-maps/marker/HighZoomMarkerContainer/hooks/useRenderStationMarker.tsx
