+++ frontend/src/components/google-maps/marker/HighZoomMarkerContainer/index.ts
@@ -1,0 +1,3 @@
+import HighZoomMarkerContainer from './HighZoomMarkerContainer';
+
+export default HighZoomMarkerContainer;
