+++ frontend/src/components/google-maps/marker/MarkersContainers.tsx
@@ -4,11 +4,11 @@ import { useExternalValue } from '@utils/external-state';
 
 import { zoomStore } from '@stores/google-maps/zoomStore';
 
+import HighZoomMarkerContainer from './HighZoomMarkerContainer';
 import RegionCountMarkersContainer from './RegionCountMarkersContainer';
 import ServerClusteredMarkersContainer from './ServerClusteredMarkersContainer';
-import StationMarkersContainer from './StationMarkersContainer';
 
-const MemoizedStationMarkersContainer = memo(StationMarkersContainer);
+const MemoizedHighZoomMarkerContainer = memo(HighZoomMarkerContainer);
 const MemoizedServerClusteredMarkersContainer = memo(ServerClusteredMarkersContainer);
 const MemoizedRegionCountMarkersContainer = memo(RegionCountMarkersContainer);
 
@@ -17,7 +17,7 @@ const MarkersContainers = () => {
 
   return (
     <>
-      {markerMode.state === 'high' && <MemoizedStationMarkersContainer />}
+      {markerMode.state === 'high' && <MemoizedHighZoomMarkerContainer />}
       {/* 이 아래는 앞으로 추가될 기능을 미리 대응하는 컴포넌트 */}
       {markerMode.state === 'middle' && <MemoizedServerClusteredMarkersContainer />}
       {markerMode.state === 'low' && <MemoizedRegionCountMarkersContainer />}
