--- frontend/src/components/google-maps/marker/StationMarkersContainer/index.ts
@@ -1,3 +1,0 @@
-import StationMarkersContainer from './StationMarkersContainer';
-
-export default StationMarkersContainer;
