+++ frontend/src/components/ui/StationSearchWindow/hooks/useStationSearchWindow.tsx
@@ -1,9 +1,9 @@
-import type { FormEvent, MouseEvent, FocusEvent, ChangeEvent } from 'react';
+import type { ChangeEvent, FocusEvent, FormEvent, MouseEvent } from 'react';
 import { useState } from 'react';
 
 import { useQueryClient } from '@tanstack/react-query';
 
-import { useRenderStationMarker } from '@marker/StationMarkersContainer/hooks/useRenderStationMarker';
+import { useRenderStationMarker } from '@marker/HighZoomMarkerContainer/hooks/useRenderStationMarker';
 
 import { useSetExternalState } from '@utils/external-state';
